Extract attribute-service registration into a reusable AttributeServices seam
AttributeServicesRegistrar owns the target-to-definition logic and
AttributeTargetsProvider is the single merged view over vendor/attributes.php
and (soon) targets discovered in attributeServicesDirectories. No behavior
change; DiscoveredAttributeTargets is empty until discovery lands.
