+++ src/DependencyInjection/AttributeServices/AttributeServicesRegistrar.php
@@ -1,0 +1,262 @@
+<?php declare(strict_types = 1);
+
+namespace PHPStan\DependencyInjection\AttributeServices;
+
+use Nette\DI\ContainerBuilder;
+use Nette\DI\Definitions\Reference;
+use Nette\DI\Definitions\ServiceDefinition;
+use Nette\DI\Definitions\Statement;
+use Nette\DI\Helpers;
+use Nette\PhpGenerator\Dumper;
+use Nette\PhpGenerator\PhpLiteral;
+use Nette\Utils\Strings;
+use olvlvl\ComposerAttributeCollector\TargetMethodParameter;
+use PHPStan\Collectors\RegistryFactory;
+use PHPStan\DependencyInjection\AutowiredParameter;
+use PHPStan\DependencyInjection\AutowiredService;
+use PHPStan\DependencyInjection\GenerateFactory;
+use PHPStan\DependencyInjection\NonAutowiredService;
+use PHPStan\DependencyInjection\RegisteredCollector;
+use PHPStan\DependencyInjection\RegisteredRule;
+use PHPStan\DependencyInjection\ValidateServiceTagsExtension;
+use PHPStan\Rules\LazyRegistry;
+use PHPStan\ShouldNotHappenException;
+use ReflectionClass;
+use function array_key_exists;
+use function array_slice;
+use function count;
+use function explode;
+use function implode;
+use function is_array;
+use function preg_split;
+use function sprintf;
+use function strcasecmp;
+use function strtolower;
+use function substr;
+use const PREG_SPLIT_DELIM_CAPTURE;
+
+/**
+ * Turns attribute targets into service definitions. Serves both sources of targets -
+ * PHPStan's own vendor/attributes.php and the directories listed
+ * in the `attributeServicesDirectories` section - through AttributeTargetsProvider.
+ */
+final class AttributeServicesRegistrar
+{
+
+	public static function registerServices(ContainerBuilder $builder, AttributeTargetsProvider $targets, ?int $level): void
+	{
+		$constructorParameters = self::collectConstructorParameters($targets);
+
+		foreach ($targets->findTargetClasses(AutowiredService::class) as $class) {
+			$reflection = new ReflectionClass($class->name);
+			$attribute = $class->attribute;
+
+			$definition = $builder->addDefinition($attribute->name)
+				->setType($class->name)
+				->setAutowired($attribute->as);
+
+			if ($attribute->factory !== null) {
+				[$ref, $method] = explode('::', $attribute->factory);
+				$definition->setFactory(new Statement([new Reference(substr($ref, 1)), $method]));
+			}
+
+			self::processConstructorParameters($builder, $class->name, $definition, $constructorParameters);
+
+			if (!$attribute->autoTag) {
+				continue;
+			}
+
+			foreach (ValidateServiceTagsExtension::getInterfaceTagMapping() as $interface => $tag) {
+				if (!$reflection->implementsInterface($interface)) {
+					continue;
+				}
+
+				$definition->addTag($tag);
+			}
+		}
+
+		foreach ($targets->findTargetClasses(NonAutowiredService::class) as $class) {
+			$attribute = $class->attribute;
+
+			$definition = $builder->addDefinition($attribute->name)
+				->setType($class->name)
+				->setAutowired(false);
+
+			if ($attribute->factory !== null) {
+				[$ref, $method] = explode('::', $attribute->factory);
+				$definition->setFactory(new Statement([new Reference(substr($ref, 1)), $method]));
+			}
+
+			self::processConstructorParameters($builder, $class->name, $definition, $constructorParameters);
+		}
+
+		foreach ($targets->findTargetClasses(GenerateFactory::class) as $class) {
+			$attribute = $class->attribute;
+			$definition = $builder->addFactoryDefinition(null)
+				->setImplement($attribute->interface);
+
+			if ($attribute->resultType !== null) {
+				$definition->getResultDefinition()->setType($attribute->resultType);
+			}
+
+			$resultDefinition = $definition->getResultDefinition();
+			self::processConstructorParameters($builder, $class->name, $resultDefinition, $constructorParameters);
+		}
+
+		if ($level === null) {
+			return;
+		}
+
+		foreach ($targets->findTargetClasses(RegisteredRule::class) as $class) {
+			$attribute = $class->attribute;
+			if ($attribute->level > $level) {
+				continue;
+			}
+
+			$definition = $builder->addDefinition(null)
+				->setFactory($class->name)
+				->setAutowired($class->name)
+				->addTag(LazyRegistry::RULE_TAG);
+
+			self::processConstructorParameters($builder, $class->name, $definition, $constructorParameters);
+		}
+
+		foreach ($targets->findTargetClasses(RegisteredCollector::class) as $class) {
+			$attribute = $class->attribute;
+			if ($attribute->level > $level) {
+				continue;
+			}
+
+			$definition = $builder->addDefinition(null)
+				->setFactory($class->name)
+				->setAutowired($class->name)
+				->addTag(RegistryFactory::COLLECTOR_TAG);
+
+			self::processConstructorParameters($builder, $class->name, $definition, $constructorParameters);
+		}
+	}
+
+	/**
+	 * @return array<lowercase-string, non-empty-list<TargetMethodParameter<AutowiredParameter>>>
+	 */
+	public static function collectConstructorParameters(AttributeTargetsProvider $targets): array
+	{
+		$constructorParameters = [];
+		foreach ($targets->findTargetMethodParameters(AutowiredParameter::class) as $parameter) {
+			if (strcasecmp($parameter->method, '__construct') !== 0) {
+				continue;
+			}
+			$lowerClass = strtolower($parameter->class);
+			$constructorParameters[$lowerClass] ??= [];
+			$constructorParameters[$lowerClass][] = $parameter;
+		}
+
+		return $constructorParameters;
+	}
+
+	/**
+	 * @param class-string $className
+	 * @param array<lowercase-string, non-empty-list<TargetMethodParameter<AutowiredParameter>>> $constructorParameters
+	 */
+	public static function processConstructorParameters(ContainerBuilder $builder, string $className, ServiceDefinition $definition, array $constructorParameters): void
+	{
+		foreach ($constructorParameters[strtolower($className)] ?? [] as $autowiredParameter) {
+			$ref = $autowiredParameter->attribute->ref;
+			if ($ref === null) {
+				$argument = self::createDeferredParameter($builder, '%' . Helpers::escape($autowiredParameter->name) . '%');
+			} elseif (Strings::match($ref, '#^@[\w\\\\]+$#D') !== null) {
+				$argument = new Reference(substr($ref, 1));
+			} else {
+				$argument = self::createDeferredParameter($builder, $ref);
+			}
+			$definition->setArgument($autowiredParameter->name, $argument);
+		}
+	}
+
+	/**
+	 * Turns a `%foo%` reference into a deferred `$this->getParameter('foo')` lookup instead of reading
+	 * ContainerBuilder::$parameters right now. Extensions registered after this one still rewrite the
+	 * parameters during loadConfiguration() - ValidateExcludePathsExtension unwraps OptionalPath objects
+	 * in `excludePaths` - and a value snapshotted here would keep the pre-rewrite contents.
+	 *
+	 * `%foo.bar%` becomes `$this->getParameter('foo')['bar']` and `%foo%/suffix` becomes a concatenation,
+	 * mirroring how Nette itself compiles references to dynamic parameters.
+	 *
+	 * @return PhpLiteral|string
+	 * @throws ShouldNotHappenException when the reference points at a parameter that does not exist
+	 */
+	private static function createDeferredParameter(ContainerBuilder $builder, string $ref)
+	{
+		$parts = preg_split('#%([\w.-]*)%#', $ref, flags: PREG_SPLIT_DELIM_CAPTURE);
+		if ($parts === false) {
+			throw new ShouldNotHappenException();
+		}
+
+		$dumper = new Dumper();
+		$lookups = [];
+		$pieces = [];
+		$withoutReferences = '';
+		foreach ($parts as $i => $part) {
+			if ($i % 2 === 0) {
+				if ($part !== '') {
+					$pieces[] = $dumper->dump($part);
+					$withoutReferences .= $part;
+				}
+				continue;
+			}
+
+			if ($part === '') {
+				// '%%' is an escaped percent sign
+				$pieces[] = $dumper->dump('%');
+				$withoutReferences .= '%';
+				continue;
+			}
+
+			$keys = explode('.', $part);
+			self::checkParameterExists($builder, $part, $keys);
+
+			$code = $dumper->format('$this->getParameter(?)', $keys[0]);
+			foreach (array_slice($keys, 1) as $key) {
+				$code .= sprintf('[%s]', $dumper->dump($key));
+			}
+
+			$lookups[] = $code;
+			$pieces[] = sprintf('(%s)', $code);
+		}
+
+		if (count($lookups) === 0) {
+			return $withoutReferences;
+		}
+
+		if (count($pieces) === 1) {
+			// the reference is the whole value, no string coercion
+			return ContainerBuilder::literal($lookups[0]);
+		}
+
+		return ContainerBuilder::literal(implode(' . ', $pieces));
+	}
+
+	/**
+	 * The values are resolved at runtime but the keys never change after loadConfiguration(),
+	 * so a typo in a reference is still caught while compiling the container.
+	 *
+	 * @param non-empty-list<string> $keys
+	 * @throws ShouldNotHappenException
+	 */
+	private static function checkParameterExists(ContainerBuilder $builder, string $ref, array $keys): void
+	{
+		$value = $builder->parameters;
+		foreach ($keys as $key) {
+			if (!is_array($value)) {
+				// a dynamic parameter - cannot be traversed while compiling
+				return;
+			}
+			if (!array_key_exists($key, $value)) {
+				throw new ShouldNotHappenException(sprintf("Missing parameter '%s'.", $ref));
+			}
+
+			$value = $value[$key];
+		}
+	}
+
+}
